requirements.txt
loguru
pyrogram
tgcrypto
python-dotenv
peewee
psycopg2-binary